Put Event Handling functions to another file
